move lightgbm up one dir
